Refactor handlers - GetElementAttribute Remove - GetElementContentSize
